host-feature-delete - fix option duplication
The short option -v was allowed for --verbose or --value.
It seems the short variant is never used within Tapper,
and the host-feature mechanism is QUITE FOR SURE not
used by anyone else, and even if so it would be very
uncommon to use -v for something other than verbose,
and if so, the user would have seen the warning and
reported it.